server/pyproject.toml
[project]
name = "okita-server"
version = "0.1.0"
description = "OKITA local backend"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlmodel>=0.0.22",
  "pydantic>=2.9",
  "httpx>=0.27",
  "apscheduler>=3.10",
  "python-jose[cryptography]>=3.3",
  "bcrypt>=4.1",
  "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.24"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
